[aptos-rest-client] Rename and move Errors to be more appropriate
