+++ crates/aptos-rest-client/src/faucet.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{error::Error, Client, Result};
+use crate::{error::FaucetClientError, Client, Result};
 use aptos_types::transaction::SignedTransaction;
 use move_deps::move_core_types::account_address::AccountAddress;
 use reqwest::{Client as ReqwestClient, Url};
@@ -48,20 +48,26 @@ impl FaucetClient {
         let query = format!("auth_key={}&amount=0&return_txns=true", address);
         url.set_query(Some(&query));
 
-        let response = self.inner.post(url).send().await.map_err(Error::request)?;
+        let response = self
+            .inner
+            .post(url)
+            .send()
+            .await
+            .map_err(FaucetClientError::request)?;
         let status_code = response.status();
-        let body = response.text().await.map_err(Error::decode)?;
+        let body = response.text().await.map_err(FaucetClientError::decode)?;
         if !status_code.is_success() {
             return Err(anyhow::anyhow!("body: {}", body));
         }
 
-        let bytes = hex::decode(body).map_err(Error::decode)?;
-        let txns: Vec<SignedTransaction> = bcs::from_bytes(&bytes).map_err(Error::decode)?;
+        let bytes = hex::decode(body).map_err(FaucetClientError::decode)?;
+        let txns: Vec<SignedTransaction> =
+            bcs::from_bytes(&bytes).map_err(FaucetClientError::decode)?;
 
         self.rest_client
             .wait_for_signed_transaction(&txns[0])
             .await
-            .map_err(Error::unknown)?;
+            .map_err(FaucetClientError::unknown)?;
 
         Ok(())
     }
@@ -75,20 +81,26 @@ impl FaucetClient {
 
         // Faucet returns the transaction that creates the account and needs to be waited on before
         // returning.
-        let response = self.inner.post(url).send().await.map_err(Error::request)?;
+        let response = self
+            .inner
+            .post(url)
+            .send()
+            .await
+            .map_err(FaucetClientError::request)?;
         let status_code = response.status();
-        let body = response.text().await.map_err(Error::decode)?;
+        let body = response.text().await.map_err(FaucetClientError::decode)?;
         if !status_code.is_success() {
-            return Err(Error::status(status_code.as_u16()).into());
+            return Err(FaucetClientError::status(status_code.as_u16()).into());
         }
 
-        let bytes = hex::decode(body).map_err(Error::decode)?;
-        let txns: Vec<SignedTransaction> = bcs::from_bytes(&bytes).map_err(Error::decode)?;
+        let bytes = hex::decode(body).map_err(FaucetClientError::decode)?;
+        let txns: Vec<SignedTransaction> =
+            bcs::from_bytes(&bytes).map_err(FaucetClientError::decode)?;
 
         self.rest_client
             .wait_for_signed_transaction(&txns[0])
             .await
-            .map_err(Error::unknown)?;
+            .map_err(FaucetClientError::unknown)?;
 
         Ok(())
     }
