+++ crates/aptos/src/common/types.rs
@@ -20,7 +20,7 @@ use aptos_crypto::{
 };
 use aptos_keygen::KeyGen;
 use aptos_rest_client::aptos_api_types::HashValue;
-use aptos_rest_client::response::RestError;
+use aptos_rest_client::error::RestError;
 use aptos_rest_client::{Client, Transaction};
 use aptos_sdk::{
     move_types::{
